refactor(forma/plot): reorganise Plot/ namespace structure
Extract series_by_role, data_range, apply_auto_scale, palette_color, and
background into PlotSpec.hpp/cpp, breaking the circular include between
SeriesBuilder.cpp and Plot.hpp. Rename SeriesBuilder to Series and PlotSource
to Source; class names are now internal detail, entry points are Plot::series()
and Plot::source() in Plot.hpp. Remove entry point free functions from
SeriesBuilder.hpp.
